Basic ropes on git stash command
Quick overview on what `git stash` is and some of the basic
commands to run with it.

Going with `git stash pop` as the default means of
extracting from the stack instead of `git stash apply`.
Former feels like the path of least surprise by taking
out what you've just put in.

As requested in #36